agents/agentic-exercise/agentic-exercise/requirements.txt
python-dotenv

langchain
langchain-core
langchain-community
langchain-experimental

langchain-ollama
langchain-tavily
langgraph-checkpoint-sqlite

langgraph-prebuilt
langgraph

chromadb
pydantic
tabulate